separate losses and conduits into their own modules
