+++ floodmodeller_api/units/structures.py
@@ -32,294 +32,6 @@ from .helpers import (
 )
 from floodmodeller_api.validation import _validate_unit
 
-
-class CULVERT(Unit):
-    """The CULVERT class supports two culvert sub-types in Flood Modeller: INLET and OUTLET. Each of these sub-types forms
-    a unique instance of the class which is differentiated by the `CULVERT.subtype` attribute. Physical culvert types (e.g. rectangualr, circular etc)
-    are provided under the CONDUIT class for consistency with Flood Modeller.
-
-    **Common attributes**
-
-    Args:
-        name (str): Unit name and upstream node label
-        ds_label (str): Downstream node label
-        us_remote_label (str): Upstream remote node label
-        ds_remote_label (str): Downstream remote node label
-        comment (str): Comment
-        loss_coefficient (float): Loss coefficient.  Outlet (default = 1.0), INLET (Trash screen head loss coefficient, default = 1.5)
-        headloss_type (str): Keyword TOTAL to denote headloss based on total head, otherwise (keyword STATIC or blank) headloss is based on static head
-        reverse_flow_mode (str): Reverse Flow Mode; keyword ZERO (for zero headloss in reverse flow) or CALCULATED (for calculated head loss in reverse flow)
-
-    **Inlet Loss Type (``CULVERT.subtype == 'INLET'``)**
-
-    Args:
-        type_code (str): options are 'Type A', 'Type B','Type C'
-        k (float): Unsubmerged inlet control loss coefficient
-        m (float): Exponent of Flow Intensity for inlet control
-        c (float): Submerged inlet control loss coefficient
-        y (float): Submerged inlet control adjustment factor
-        ki (float): Outlet control loss coefficient
-        screen_width (float): Trash screen width (m)
-        bar_proportion (float): Proportion of trash screen area occupied by bars (0 to 1.0)
-        debris_proportion (float): Blockage ratio (proportion of trash screen area occupied by debris) (0 to 1.0)
-        max_screen_height (float): Max. Trash Screen Height *see Flood Modeller help for further information
-
-    **Outlet Loss Type (``CULVERT.subtype == 'OUTLET'``)**
-
-    No additional attributes required for OUTLET subtype
-
-    Returns:
-        CULVERT: Flood Modeller CULVERT unit class object
-    """
-
-    _unit = "CULVERT"
-
-    def _read(self, block):
-        """Function to read a given CULVERT block and store data as class attributes"""
-
-        # Extract common attributes
-        self._subtype = block[1].split(" ")[0].strip()
-        self.comment = block[0].replace("CULVERT", "").strip()
-        labels = split_n_char(f"{block[2]:<{4*self._label_len}}", self._label_len)
-        self.name = labels[0]
-        self.ds_label = labels[1]
-        self.us_remote_label = labels[2]
-        self.ds_remote_label = labels[3]
-
-        # Extract subtype specific attributes
-
-        if self.subtype == "INLET":
-            # Defaults applied are equivellent to those provide if Culvert Wizard cancelled.
-
-            # Read first set of general parameters
-            params = split_10_char(f"{block[3]:<60}")
-            self.k = _to_float(params[0], 0.0)
-            self.m = _to_float(params[1], 0.0)
-            self.c = _to_float(params[2], 0.0)
-            self.y = _to_float(params[3], 0.0)
-            self.ki = _to_float(params[4], 0.0)
-            self.type_code = _to_str(params[5], "A")
-
-            # Read trash screen and remaining general parameters
-            params1 = split_10_char(f"{block[4]:<70}")
-            self.screen_width = _to_float(params1[0], 0.0)
-            self.bar_proportion = _to_float(params1[1], 0.0)
-            self.debris_proportion = _to_float(params1[2], 0.0)
-            self.loss_coefficient = _to_float(params1[3], 0.0)
-            self.reverse_flow_mode = _to_str(params1[4], "CALCULATED", check_float=True)
-            self.headloss_type = _to_str(params1[5], "TOTAL")
-            self.max_screen_height = _to_float(params1[6], 0.0)
-
-        elif self.subtype == "OUTLET":
-            params = split_10_char(f"{block[3]:<30}")
-            self.loss_coefficient = _to_float(params[0], 1.0)
-            self.reverse_flow_mode = _to_str(params[1], "CALCULATED")
-            self.headloss_type = _to_str(params[2], "TOTAL")
-
-        else:
-            # This else block is triggered for culvert subtypes which aren't yet supported, and just keeps the '_block' in it's raw state to write back.
-            print(
-                f'This Culvert sub-type: "{self.subtype}" is currently unsupported for reading/editing'
-            )
-            self._raw_block = block
-
-        # TODO: Create from blank.  Not supported currently as CULVERT has multiple subtypes
-
-    def _write(self):
-        """Function to write a valid CULVERT block"""
-
-        _validate_unit(self)
-
-        header = "CULVERT " + self.comment
-        labels = join_n_char_ljust(
-            self._label_len,
-            self.name,
-            self.ds_label,
-            self.us_remote_label,
-            self.ds_remote_label,
-        )
-        c_block = [header, self.subtype, labels]
-
-        if self.subtype == "INLET":
-            params = join_10_char(
-                self.k, self.m, self.c, self.y, self.ki, self.type_code, dp=4
-            )
-            params1 = join_10_char(
-                self.screen_width,
-                self.bar_proportion,
-                self.debris_proportion,
-                self.loss_coefficient,
-                self.reverse_flow_mode,
-                self.headloss_type,
-                self.max_screen_height,
-                dp=4,
-            )
-
-            c_block.extend([params, params1])
-
-            return c_block
-
-        elif self.subtype == "OUTLET":
-            params = join_10_char(
-                self.loss_coefficient, self.reverse_flow_mode, self.headloss_type
-            )
-
-            c_block.append(params)
-            return c_block
-
-        else:
-            return self._raw_block
-
-
-class BLOCKAGE(Unit):
-    """Class to hold and process BLOCKAGE unit type.
-
-    Args:
-        comment (str): Comment included in unit.
-        name (str): Upstream label name
-        ds_label (str): Downstream label
-        us_reference_label (str): Upstream reference label
-        ds_reference_label (str): Downstream reference label
-        constriction_label (str): Constriction reference label
-        inlet_loss (float): Inlet loss coefficient
-        outlet_loss (float): Outlet loss coefficient
-        timeoffset (float): Time Datum Adjustment
-        timeunit_blockage (str): Unit of time, e.g. ‘HOURS’, ‘MINUTES’ or ‘SECONDS’. See Flood Modeller documentation for all available options.
-        extendmethod (str): Data extending method: ‘EXTEND’, ‘NOEXTEND’ or ‘REPEAT’. Defaults to None.
-        data (pandas.Series): Series object with variable ``'blockage'`` and index ``'Time'``. Defaults to None.
-
-
-    Returns:
-        BLOCKAGE: Flood Modeller BLOCAKGE Unit class object
-    """
-
-    _unit = "BLOCKAGE"
-
-    def _read(self, block):
-        """Function to read a given BLOCKAGE block and store data as class attributes"""
-
-        # Extract comment and revision number
-        b = block[0].replace("BLOCKAGE #revision#", " ").strip()
-        self._revision = _to_int(b[0], 1)
-        self.comment = b[1:].strip()
-
-        # Extract labels
-        labels = split_n_char(f"{block[1]:<{5*self._label_len}}", self._label_len)
-        self.name = labels[0]
-        self.ds_label = labels[1]
-        self.us_reference_label = labels[2]
-        self.ds_reference_label = labels[3]
-        self.constriction_label = labels[4]
-
-        # Extract inlet and outlet loss coefficients
-        params = split_10_char(f"{block[2]:<20}")
-        self.inlet_loss = _to_float(params[0], 1.5)
-        self.outlet_loss = _to_float(params[1], 1.0)
-
-        # Extract blockage timeseries parameters
-        params1 = split_10_char(f"{block[3]:<40}")
-        self.nrows = int(params1[0])
-        self.timeoffset = _to_float(params1[1])
-
-        self.timeunit = _to_str(params1[2], "HOURS", check_float=True)
-        if self.timeunit == "DATE":
-            self.timeunit = "DATES"  # Parameter value updated to 'DATES' for consistency with other unit types.  'DATE' and 'DATES' both accepted for blockage unit ONLY
-
-        self.extendmethod = _to_str(params1[3], "NOEXTEND")
-
-        # Extract blockage to timeseries
-        data_list = (
-            _to_data_list(block[4:], num_cols=2, date_col=0)
-            if self.timeunit == "DATES"
-            else _to_data_list(block[4:], num_cols=2)
-        )  # Enforced two columns as Flood Modeller saves old parameters when using DATES (also to avoid extra 'HOURS' bug)
-
-        self.data = pd.DataFrame(data_list, columns=["Time", "Blockage"])
-        self.data = self.data.set_index("Time")
-        self.data = self.data["Blockage"]
-
-    def _write(self):
-        """Function to write a valid BLOCKAGE block"""
-
-        _validate_unit(self)
-
-        # Custom validation for blockage percentage
-        if self.data.max() > 1 or self.data.min() < 0:
-            raise ValueError(
-                f"Parameter error with {self.__repr__()} - blockage percentage must be between 0 and 1"
-            )
-
-        header = f"BLOCKAGE #revision#{self._revision} {self.comment}"
-        labels = join_n_char_ljust(
-            self._label_len,
-            self.name,
-            self.ds_label,
-            self.us_reference_label,
-            self.ds_reference_label,
-            self.constriction_label,
-        )
-        params = join_10_char(self.inlet_loss, self.outlet_loss)
-        self.nrows = len(self.data)
-        params1 = join_10_char(
-            self.nrows, self.timeoffset, self.timeunit, self.extendmethod
-        )
-
-        blockage_block = [header, labels, params, params1]
-
-        if self.timeunit == "DATES":
-            blockage_data = [
-                f"{t:<20}{join_10_char(b)}" for t, b in self.data.iteritems()
-            ]
-        else:
-            blockage_data = [join_10_char(t, b) for t, b in self.data.iteritems()]
-
-        blockage_block.extend(blockage_data)
-
-        return blockage_block
-
-    def _create_from_blank(
-        self,
-        name="new_blockage",
-        _revision=1,
-        comment="",
-        ds_label="",
-        us_reference_label="",
-        ds_reference_label="",
-        constriction_label="",
-        inlet_loss=1.5,
-        outlet_loss=1.0,
-        timeoffset=0.0,
-        timeunit="HOURS",
-        extendmethod="EXTEND",
-        data=None,
-    ):
-
-        # Initiate new BLOCKAGE unit
-        for param, val in {
-            "name": name,
-            "_revision": _revision,
-            "comment": comment,
-            "ds_label": ds_label,
-            "us_reference_label": us_reference_label,
-            "ds_reference_label": ds_reference_label,
-            "constriction_label": constriction_label,
-            "inlet_loss": inlet_loss,
-            "outlet_loss": outlet_loss,
-            "timeoffset": timeoffset,
-            "timeunit": timeunit,
-            "timeunit": timeunit,
-            "extendmethod": extendmethod,
-            "data": data,
-        }.items():
-            setattr(self, param, val)
-
-        self.data = (
-            data
-            if isinstance(data, pd.Series)
-            else pd.Series([0.0], index=[0.0], name="Blockage")
-        )
-
-
 class BRIDGE(Unit):
     """Class to hold and process BRIDGE unit type. The Bridge class supports the three main bridge sub-types in
     Flood Modeller: Arch, USBPR1978 and Pierloss. Each of these sub-types forms a unique instance of the class
@@ -812,175 +524,6 @@ class BRIDGE(Unit):
 
         else:
             return self._raw_block
-
-
-class CONDUIT(Unit):
-    """The Conduit class supports two conduit sub-types in Flood Modeller: RECTANGULAR and CIRCULAR. Each of these sub-types forms
-    a unique instance of the class which is differentiated by the `CONDUIT.subtype` attribute. All conduit types have the same common
-    attributes:
-
-    **Common Attributes**
-
-    Args:
-        name (str): Conduit section name
-        spill (str): Spill label
-        comment (str): Comment included in unit
-        dist_to_next (float): Distance to next section in metres
-        subtype (str): Defines the type of conduit unit (*Should not be changed*)
-
-    **Rectangular Type (``CONDUIT.subtype == 'RECTANGULAR'``)**
-
-    Args:
-        friction_eq (str): Friction equation to use (``'MANNING'`` or ``'COLEBROOK-WHITE'``)
-        invert (float): Elevation of invert above datum (m)
-        width (float): Width of conduit (m)
-        height (float): Height of conduit (m)
-        use_bottom_slot (str): Whether to include bottom slot (``'ON'``, ``'OFF'`` or ``'GLOBAL'``). Setting it to 'GLOBAL' will use the default option specified in IEF.
-        bottom_slot_dist (float): Distance of slot top above invert (m)
-        bottom_slot_depth (float): Total depth of bottom slot (m)
-        use_top_slot (str): Whether to include top slot (``'ON'``, ``'OFF'`` or ``'GLOBAL'``). Setting it to 'GLOBAL' will use the default option specified in IEF.
-        top_slot_dist (float): Distance of slot bottom below soffit (m)
-        top_slot_depth (float): Total depth of top slot (m)
-        friction_on_invert (float): Friction value for conduit invert
-        friction_on_walls (float): Friction value for conduit walls
-        friction_on_soffit (float): Friction value for conduit soffit
-
-    **Circular Type (``CONDUIT.subtype == 'CIRCULAR'``)**
-
-    Args:
-        friction_eq (str): Friction equation to use (``'MANNING'`` or ``'COLEBROOK-WHITE'``)
-        invert (float): Elevation of invert above datum (m)
-        diameter (float): Diameter of conduit (m)
-        use_bottom_slot (str): Whether to include bottom slot (``'ON'``, ``'OFF'`` or ``'GLOBAL'``). Setting it to 'GLOBAL' will use the default option specified in IEF.
-        bottom_slot_dist (float): Distance of slot top above invert (m)
-        bottom_slot_depth (float): Total depth of bottom slot (m)
-        use_top_slot (str): Whether to include top slot (``'ON'``, ``'OFF'`` or ``'GLOBAL'``). Setting it to 'GLOBAL' will use the default option specified in IEF.
-        top_slot_dist (float): Distance of slot bottom below soffit (m)
-        top_slot_depth (float): Total depth of top slot (m)
-        friction_below_axis (float): Friction value for conduit below axis
-        friction_above_axis (float): Friction value for conduit above axis
-
-
-    Raises:
-        NotImplementedError: Raised if class is initialised without existing Conduit block (i.e. if attempting to create new
-            Conduit unit). This will be an option for future releases
-
-    Returns:
-        CONDUIT: Flood Modeller CONDUIT Unit class object
-    """
-
-    _unit = "CONDUIT"
-
-    def _read(self, c_block):
-        """Function to read a given CONDUIT block and store data as class attributes"""
-        self._subtype = c_block[1].split(" ")[0].strip()
-        # Extends label line to be correct length before splitting to pick up blank labels
-        labels = split_n_char(f"{c_block[2]:<{2*self._label_len}}", self._label_len)
-        self.name = labels[0]
-        self.spill = labels[1]
-        self.comment = c_block[0].replace("CONDUIT", "").strip()
-
-        # Read CIRCULAR type unit
-        if self.subtype == "CIRCULAR":
-            # Read Params
-            self.dist_to_next = _to_float(split_10_char(c_block[3])[0])
-            self.friction_eq = c_block[4].strip()
-            params = split_10_char(f"{c_block[5]:<80}")
-            self.invert = _to_float(params[0])
-            self.diameter = _to_float(params[1])
-            self.use_bottom_slot = _to_str(params[2], "GLOBAL")
-            self.bottom_slot_dist = _to_float(params[3])
-            self.bottom_slot_depth = _to_float(params[4])
-            self.use_top_slot = _to_str(params[5], "GLOBAL")
-            self.top_slot_dist = _to_float(params[6])
-            self.top_slot_depth = _to_float(params[7])
-            friction_params = split_10_char(f"{c_block[6]:<20}")
-            self.friction_below_axis = _to_float(friction_params[0])
-            self.friction_above_axis = _to_float(friction_params[1])
-
-        elif self.subtype == "RECTANGULAR":
-            # Read Params
-            self.dist_to_next = _to_float(split_10_char(c_block[3])[0])
-            self.friction_eq = c_block[4].strip()
-            params = split_10_char(f"{c_block[5]:<90}")
-            self.invert = _to_float(params[0])
-            self.width = _to_float(params[1])
-            self.height = _to_float(params[2])
-            self.use_bottom_slot = _to_str(params[3], "GLOBAL")
-            self.bottom_slot_dist = _to_float(params[4])
-            self.bottom_slot_depth = _to_float(params[5])
-            self.use_top_slot = _to_str(params[6], "GLOBAL")
-            self.top_slot_dist = _to_float(params[7])
-            self.top_slot_depth = _to_float(params[8])
-            friction_params = split_10_char(f"{c_block[6]:<30}")
-            self.friction_on_invert = _to_float(friction_params[0])
-            self.friction_on_walls = _to_float(friction_params[1])
-            self.friction_on_soffit = _to_float(friction_params[2])
-
-        else:
-            # This else block is triggered for conduit subtypes which aren't yet supported, and just keeps the '_block' in it's raw state to write back.
-            print(
-                f'This Conduit sub-type: "{self.subtype}" is currently unsupported for reading/editing'
-            )
-            self._raw_block = c_block
-
-    def _write(self):
-        """Function to write a valid CONDUIT block"""
-        _validate_unit(self)  # Function to check the params are valid for CONDUIT unit
-        header = "CONDUIT " + self.comment
-        labels = join_n_char_ljust(self._label_len, self.name, self.spill)
-        c_block = [header, self.subtype, labels]
-
-        if self.subtype == "CIRCULAR":
-            params = join_10_char(
-                self.invert,
-                self.diameter,
-                self.use_bottom_slot,
-                self.bottom_slot_dist,
-                self.bottom_slot_depth,
-                self.use_top_slot,
-                self.top_slot_dist,
-                self.top_slot_depth,
-            )
-            friction_params = (
-                f"{self.friction_below_axis:>10.4f}{self.friction_below_axis:>10.4f}"
-            )
-            c_block.extend(
-                [
-                    f"{self.dist_to_next:>10.3f}",
-                    self.friction_eq,
-                    params,
-                    friction_params,
-                ]
-            )
-            return c_block
-
-        elif self.subtype == "RECTANGULAR":
-            params = join_10_char(
-                self.invert,
-                self.width,
-                self.height,
-                self.use_bottom_slot,
-                self.bottom_slot_dist,
-                self.bottom_slot_depth,
-                self.use_top_slot,
-                self.top_slot_dist,
-                self.top_slot_depth,
-            )
-            friction_params = f"{self.friction_on_invert:>10.4f}{self.friction_on_walls:>10.4f}{self.friction_on_soffit:>10.4f}"
-            c_block.extend(
-                [
-                    f"{self.dist_to_next:>10.3f}",
-                    self.friction_eq,
-                    params,
-                    friction_params,
-                ]
-            )
-            return c_block
-
-        else:
-            return self._raw_block
-
 
 class SLUICE(Unit):
     """The Sluice class supports two sluice sub-types in Flood Modeller: RADIAL and VERTICAL. Each of these sub-types forms
